requirements.txt
colorlog
requests
pyyaml
pyTelegramBotAPI
pytz
